requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
boto3==1.42.34
botocore==1.42.34
celery==5.3.1
celery-types==0.24.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.1
Django==6.0.1
django-debug-toolbar==5.0.1
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.2.12
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
ruff==0.14.11
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
pyTelegramBotAPI==4.15.4
